feat(KNO-4367): abstract readExtractedFileSync
