+++ src/commands/layout/pull.ts
@@ -17,8 +17,8 @@ import { promptToConfirm, spinner } from "@/lib/helpers/ux";
 import * as EmailLayout from "@/lib/marshal/email-layout";
 import { WithAnnotation } from "@/lib/marshal/shared/types";
 import {
-  ensureResourceDirForTarget,
   EmailLayoutDirContext,
+  ensureResourceDirForTarget,
   ResourceTarget,
 } from "@/lib/run-context";
