+++ src/lib/marshal/workflow/reader.ts
@@ -5,13 +5,7 @@ import * as fs from "fs-extra";
 import { hasIn, set } from "lodash";
 
 import { formatErrors, JsonDataError, SourceError } from "@/lib/helpers/error";
-import {
-  ParsedJson,
-  parseJson,
-  ParseJsonResult,
-  readJson,
-} from "@/lib/helpers/json";
-import { validateLiquidSyntax } from "@/lib/helpers/liquid";
+import { ParseJsonResult, readJson } from "@/lib/helpers/json";
 import {
   AnyObj,
   getLastFound,
@@ -19,13 +13,15 @@ import {
   ObjPath,
   omitDeep,
 } from "@/lib/helpers/object";
+import {
+  FILEPATH_MARKED_RE,
+  readExtractedFileSync,
+} from "@/lib/marshal/shared/helpers";
 import { WorkflowDirContext } from "@/lib/run-context";
 
 import {
-  FILEPATH_MARKED_RE,
   isWorkflowDir,
   lsWorkflowJson,
-  VISUAL_BLOCKS_JSON,
   WORKFLOW_JSON,
   WorkflowCommandTarget,
 } from "./helpers";
@@ -38,10 +34,6 @@ export type WorkflowDirData = WorkflowDirContext & {
 // For now we support up to two levels of content extraction in workflow.json.
 // (e.g. workflow.json, then visual_blocks.json)
 const MAX_EXTRACTION_LEVEL = 2;
-
-// The following files are exepected to have valid json content, and should be
-// decoded and joined into the main workflow.json.
-const DECODABLE_JSON_FILES = new Set([VISUAL_BLOCKS_JSON]);
 
 /*
  * Validate the file path format of an extracted field. The file path must be:
@@ -109,68 +101,6 @@ const validateExtractedFilePath = (
   uniqueFilePaths[val] = true;
 
   return undefined;
-};
-
-/*
- * Read the file at the given path if it exists, validate the content as
- * applicable, and return the content string or an error.
- */
-type ExtractedFileContent = string | ParsedJson;
-type ReadExtractedFileResult =
-  | [undefined, JsonDataError]
-  | [ExtractedFileContent, undefined];
-
-const readExtractedFileSync = (
-  relpath: string,
-  workflowDirCtx: WorkflowDirContext,
-  objPathToFieldStr = "",
-): ReadExtractedFileResult => {
-  // Check if the file actually exists at the given file path.
-  const abspath = path.resolve(workflowDirCtx.abspath, relpath);
-  const exists = fs.pathExistsSync(abspath);
-  if (!exists) {
-    const error = new JsonDataError(
-      "must be a relative path string to a file that exists",
-      objPathToFieldStr,
-    );
-    return [undefined, error];
-  }
-
-  // Read the file and check for valid liquid syntax given it is supported
-  // across all message templates and file extensions.
-  const contentStr = fs.readFileSync(abspath, "utf8");
-  const liquidParseError = validateLiquidSyntax(contentStr);
-
-  if (liquidParseError) {
-    const error = new JsonDataError(
-      `points to a file that contains invalid liquid syntax (${relpath})\n\n` +
-        formatErrors([liquidParseError], { indentBy: 2 }),
-      objPathToFieldStr,
-    );
-
-    return [undefined, error];
-  }
-
-  // If the file is expected to contain decodable json, then parse the contentStr
-  // as such.
-  const fileName = path.basename(abspath.toLowerCase());
-  const decodable = DECODABLE_JSON_FILES.has(fileName);
-
-  const [content, jsonParseErrors] = decodable
-    ? parseJson(contentStr)
-    : [contentStr, []];
-
-  if (jsonParseErrors.length > 0) {
-    const error = new JsonDataError(
-      `points to a file with invalid content (${relpath})\n\n` +
-        formatErrors(jsonParseErrors, { indentBy: 2 }),
-      objPathToFieldStr,
-    );
-
-    return [undefined, error];
-  }
-
-  return [content!, undefined];
 };
 
 /*
@@ -424,4 +354,4 @@ export const readAllForCommandTarget = async (
 };
 
 // Exported for tests.
-export { checkIfValidExtractedFilePathFormat, readExtractedFileSync };
+export { checkIfValidExtractedFilePathFormat };
